Also fix clang warning with simulated packetloss
Also implement QALIGN on msvc
Add a few range checkers and descriptions
Fix minimize command under SDL builds not respecting video recording state like win32/linux glimp does

+++ src/qcommon/net_chan.c
@@ -529,7 +529,7 @@ void NET_FlushPacketQueue( void )
 void NET_SendPacket( netsrc_t sock, int length, const void *data, const netadr_t *to ) {
 #ifndef DEDICATED
 	if ( sock == NS_CLIENT && cl_packetloss->integer > 0 ) {
-		if ( ( (float)rand() / RAND_MAX ) * 100 <= cl_packetloss->integer ) {
+		if ( ( (float)rand() / (float)RAND_MAX ) * 100 <= cl_packetloss->integer ) {
 			if ( showpackets->integer ) {
 				Com_Printf( "drop packet %4i\n", length );
 			}
@@ -538,7 +538,7 @@ void NET_SendPacket( netsrc_t sock, int length, const void *data, const netadr_t
 	}
 #endif
 	if ( sock == NS_SERVER && sv_packetloss->integer > 0 ) {
-		if ( ( (float)rand() / RAND_MAX ) * 100 <= sv_packetloss->integer ) {
+		if ( ( (float)rand() / (float)RAND_MAX ) * 100 <= sv_packetloss->integer ) {
 			if ( showpackets->integer ) {
 				Com_Printf( "drop packet %4i\n", length );
 			}
